Move server/src/routes into server/src/features/api in order to add server/src/features/crons package
