+++ server/src/domain/services/job-service.js
@@ -45,5 +45,5 @@ function getJobs() {
 
 module.exports = {
   fetchAndCacheJobs,
-  getJobs
+  getJobs,
 };
