+++ server/src/features/api/jobs.js
@@ -1,8 +1,8 @@
 const express = require('express');
 
 const router = express.Router();
-const auth = require('../infrastructure/middlewares/auth');
-const jobService = require('../domain/services/job-service');
+const auth = require('../../infrastructure/middlewares/auth');
+const jobService = require('../../domain/services/job-service');
 
 router.get('/', auth, (req, res) => {
   jobService
